Controller.KasaBulb/requirements.txt
quart
flask_cors
flask-pymongo
fastapi
requests
gevent
python-kasa
pymongo